MainViewModelの機能を複数のファイルに移動
