+++ src/Beutl/ViewModels/MainViewModel.cs
@@ -1,17 +1,12 @@
-﻿#pragma warning disable CS0436
-
-using System.Windows.Input;
+﻿using System.Windows.Input;
 
 using Avalonia;
-using Avalonia.Collections;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Input;
 using Avalonia.Input.Platform;
 
 using Beutl.Api;
 using Beutl.Api.Services;
-
-using Beutl.Configuration;
 using Beutl.Services;
 using Beutl.Services.PrimitiveImpls;
 using Beutl.Services.StartupTasks;
@@ -25,10 +20,8 @@ using Serilog;
 
 namespace Beutl.ViewModels;
 
-// Todo: StartupやMenuBarViewModelなど複数のクラスに分ける
 public sealed class MainViewModel : BasePageViewModel
 {
-    private readonly ILogger _logger;
     private readonly BeutlApiApplication _beutlClients;
     private readonly HttpClient _authorizedHttpClient;
 
@@ -53,10 +46,10 @@ public sealed class MainViewModel : BasePageViewModel
 
     public MainViewModel()
     {
-        _logger = Log.ForContext<MainViewModel>();
         _authorizedHttpClient = new HttpClient();
         _beutlClients = new BeutlApiApplication(_authorizedHttpClient);
         ExtensionProvider.Current = _beutlClients.GetResource<ExtensionProvider>();
+        MenuBar = new MenuBarViewModel();
 
         IsProjectOpened = ProjectService.Current.IsOpened;
 
@@ -68,110 +61,6 @@ public sealed class MainViewModel : BasePageViewModel
             .SelectMany(i => i?.Context ?? Observable.Empty<IEditorContext?>())
             .Select(v => v is EditViewModel);
 
-        AddToProject = new(isProjectOpenedAndTabOpened);
-        RemoveFromProject = new(isProjectOpenedAndTabOpened);
-        AddLayer = new(isSceneOpened);
-        DeleteLayer = new(isSceneOpened);
-        ExcludeLayer = new(isSceneOpened);
-        CutLayer = new(isSceneOpened);
-        CopyLayer = new(isSceneOpened);
-        PasteLayer = new(isSceneOpened);
-        CloseFile = new(EditorService.Current.SelectedTabItem.Select(i => i != null));
-        CloseProject = new(IsProjectOpened);
-        Save = new(IsProjectOpened);
-        SaveAll = new(IsProjectOpened);
-        Undo = new(IsProjectOpened);
-        Redo = new(IsProjectOpened);
-
-        Save.Subscribe(async () =>
-        {
-            EditorTabItem? item = EditorService.Current.SelectedTabItem.Value;
-            if (item != null)
-            {
-                try
-                {
-                    bool result = await (item.Commands.Value == null ? ValueTask.FromResult(false) : item.Commands.Value.OnSave());
-
-                    if (result)
-                    {
-                        NotificationService.ShowSuccess(string.Empty, string.Format(Message.ItemSaved, item.FileName));
-                    }
-                    else
-                    {
-                        Type type = item.Extension.Value.GetType();
-                        _logger.Error("{Extension} failed to save file", type.FullName ?? type.Name);
-                        NotificationService.ShowInformation(string.Empty, Message.OperationCouldNotBeExecuted);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    _logger.Error(ex, "Failed to save file");
-                    NotificationService.ShowError(string.Empty, Message.OperationCouldNotBeExecuted);
-                }
-            }
-        });
-
-        SaveAll.Subscribe(async () =>
-        {
-            Project? project = ProjectService.Current.CurrentProject.Value;
-            int itemsCount = 0;
-
-            try
-            {
-                project?.Save(project.FileName);
-                itemsCount++;
-
-                foreach (EditorTabItem? item in EditorService.Current.TabItems)
-                {
-                    if (item.Commands.Value != null)
-                    {
-                        if (await item.Commands.Value.OnSave())
-                        {
-                            itemsCount++;
-                        }
-                        else
-                        {
-                            Type type = item.Extension.Value.GetType();
-                            _logger.Error("{Extension} failed to save file", type.FullName ?? type.Name);
-                            NotificationService.ShowError(Message.Unable_to_save_file, item.FileName.Value);
-                        }
-                    }
-                }
-
-                NotificationService.ShowSuccess(string.Empty, string.Format(Message.ItemsSaved, itemsCount.ToString()));
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Failed to save files");
-                NotificationService.ShowError(string.Empty, Message.OperationCouldNotBeExecuted);
-            }
-        });
-
-        CloseFile.Subscribe(() =>
-        {
-            EditorTabItem? tabItem = EditorService.Current.SelectedTabItem.Value;
-            if (tabItem != null)
-            {
-                EditorService.Current.CloseTabItem(
-                    tabItem.FilePath.Value,
-                    tabItem.TabOpenMode);
-            }
-        });
-        CloseProject.Subscribe(() => ProjectService.Current.CloseProject());
-
-        Undo.Subscribe(async () =>
-        {
-            IKnownEditorCommands? commands = EditorService.Current.SelectedTabItem.Value?.Commands.Value;
-            if (commands != null)
-                await commands.OnUndo();
-        });
-        Redo.Subscribe(async () =>
-        {
-            IKnownEditorCommands? commands = EditorService.Current.SelectedTabItem.Value?.Commands.Value;
-            if (commands != null)
-                await commands.OnRedo();
-        });
-
         Pages = new()
         {
             new(EditPageExtension.Instance),
@@ -180,113 +69,13 @@ public sealed class MainViewModel : BasePageViewModel
         };
         SettingsPage = new(SettingsPageExtension.Instance, new SettingsPageViewModel(_beutlClients));
         SelectedPage.Value = Pages[0];
-        ViewConfig viewConfig = GlobalConfiguration.Instance.ViewConfig;
-        viewConfig.RecentFiles.ForEachItem(
-            item => RecentFileItems.Insert(0, item),
-            item => RecentFileItems.Remove(item),
-            () => RecentFileItems.Clear());
-
-        viewConfig.RecentProjects.ForEachItem(
-            item => RecentProjectItems.Insert(0, item),
-            item => RecentProjectItems.Remove(item),
-            () => RecentProjectItems.Clear());
-
-        OpenRecentFile.Subscribe(file => EditorService.Current.ActivateTabItem(file, TabOpenMode.YourSelf));
-
-        OpenRecentProject.Subscribe(file =>
-        {
-            if (!File.Exists(file))
-            {
-                NotificationService.ShowInformation("", Message.FileDoesNotExist);
-            }
-            else if (ProjectService.Current.OpenProject(file) == null)
-            {
-                NotificationService.ShowInformation("", Message.CouldNotOpenProject);
-            }
-        });
 
         KeyBindings = CreateKeyBindings();
     }
 
     public bool IsDebuggerAttached { get; } = Debugger.IsAttached;
 
-    // File
-    //    Create new
-    //       Project
-    //       File
-    //    Open
-    //       Project
-    //       File
-    //    Close
-    //    Close project
-    //    Save
-    //    Save all
-    //    Recent files
-    //    Recent projects
-    //    Exit
-    public ReactiveCommand CreateNewProject { get; } = new();
-
-    public ReactiveCommand CreateNew { get; } = new();
-
-    public ReactiveCommand OpenProject { get; } = new();
-
-    public ReactiveCommand OpenFile { get; } = new();
-
-    public ReactiveCommand CloseFile { get; }
-
-    public ReactiveCommand CloseProject { get; }
-
-    public ReactiveCommand Save { get; }
-
-    public ReactiveCommand SaveAll { get; }
-
-    public ReactiveCommand<string> OpenRecentFile { get; } = new();
-
-    public ReactiveCommand<string> OpenRecentProject { get; } = new();
-
-    public CoreList<string> RecentFileItems { get; } = new();
-
-    public CoreList<string> RecentProjectItems { get; } = new();
-
-    public ReactiveCommand Exit { get; } = new();
-
-    // Edit
-    //    Undo
-    //    Redo
-    public ReactiveCommand Undo { get; }
-
-    public ReactiveCommand Redo { get; }
-
-    // Project
-    //    Add
-    //    Remove
-    public ReactiveCommand AddToProject { get; }
-
-    public ReactiveCommand RemoveFromProject { get; }
-
-    // Scene
-    //    New
-    //    Settings
-    //    Layer
-    //       Add
-    //       Delete
-    //       Exclude
-    //       Cut
-    //       Copy
-    //       Paste
-    public ReactiveCommand NewScene { get; } = new();
-
-    public ReactiveCommand AddLayer { get; }
-
-    public ReactiveCommand DeleteLayer { get; }
-
-    public ReactiveCommand ExcludeLayer { get; }
-
-    public ReactiveCommand CutLayer { get; }
-
-    public ReactiveCommand CopyLayer { get; }
-
-    public ReactiveCommand PasteLayer { get; }
+    public MenuBarViewModel MenuBar { get; }
 
     public NavItemViewModel SettingsPage { get; }
 
@@ -301,30 +90,6 @@ public sealed class MainViewModel : BasePageViewModel
     public Task RunStartupTask()
     {
         return new Startup(_beutlClients, this).Run();
-    }
-
-    public async ValueTask<CheckForUpdatesResponse?> CheckForUpdates()
-    {
-        try
-        {
-            return await _beutlClients.App.CheckForUpdatesAsync(GitVersionInformation.NuGetVersion);
-        }
-        catch (Exception ex)
-        {
-            _logger.Error(ex, "An error occurred while checking for updates");
-            ErrorHandle(ex);
-            return null;
-        }
-    }
-
-    public ToolTabExtension[] GetToolTabExtensions()
-    {
-        return _beutlClients.GetResource<ExtensionProvider>().GetExtensions<ToolTabExtension>();
-    }
-
-    public EditorExtension[] GetEditorExtensions()
-    {
-        return _beutlClients.GetResource<ExtensionProvider>().GetExtensions<EditorExtension>();
     }
 
     public void RegisterServices()
@@ -400,30 +165,30 @@ public sealed class MainViewModel : BasePageViewModel
         var list = new List<KeyBinding>
         {
             // CreateNewProject: Ctrl+Shift+N
-            KeyBinding(Key.N, modifier | KeyModifiers.Shift, CreateNewProject),
+            KeyBinding(Key.N, modifier | KeyModifiers.Shift, MenuBar.CreateNewProject),
             // CreateNew: Ctrl+N
-            KeyBinding(Key.N, modifier, CreateNew),
+            KeyBinding(Key.N, modifier, MenuBar.CreateNew),
             // OpenProject: Ctrl+Shift+O
-            KeyBinding(Key.O, modifier | KeyModifiers.Shift, OpenProject),
+            KeyBinding(Key.O, modifier | KeyModifiers.Shift, MenuBar.OpenProject),
             // OpenFile: Ctrl+O
-            KeyBinding(Key.O, modifier, OpenFile),
+            KeyBinding(Key.O, modifier, MenuBar.OpenFile),
             // Save: Ctrl+S
-            KeyBinding(Key.S, modifier, Save),
+            KeyBinding(Key.S, modifier, MenuBar.Save),
             // SaveAll: Ctrl+Shift+S
-            KeyBinding(Key.S, modifier | KeyModifiers.Shift, SaveAll),
+            KeyBinding(Key.S, modifier | KeyModifiers.Shift, MenuBar.SaveAll),
             // Exit: Alt+F4
-            KeyBinding(Key.F4, KeyModifiers.Alt, Exit),
+            KeyBinding(Key.F4, KeyModifiers.Alt, MenuBar.Exit),
         };
 
         if (config != null)
         {
-            list.AddRange(config.Undo.Select(x => KeyBinding(x.Key, x.KeyModifiers, Undo)));
-            list.AddRange(config.Redo.Select(x => KeyBinding(x.Key, x.KeyModifiers, Redo)));
+            list.AddRange(config.Undo.Select(x => KeyBinding(x.Key, x.KeyModifiers, MenuBar.Undo)));
+            list.AddRange(config.Redo.Select(x => KeyBinding(x.Key, x.KeyModifiers, MenuBar.Redo)));
         }
         else
         {
-            list.Add(KeyBinding(Key.Z, modifier, Undo));
-            list.Add(KeyBinding(Key.R, modifier, Redo));
+            list.Add(KeyBinding(Key.Z, modifier, MenuBar.Undo));
+            list.Add(KeyBinding(Key.R, modifier, MenuBar.Redo));
         }
 
         return list;
